refactor(types): split type files
