+++ src/types/orders.go
@@ -1,0 +1,8 @@
+package types
+
+import "Driver-go/elevio"
+
+type DirnBehaviourPair struct {
+	Dirn      elevio.MotorDirection
+	Behaviour ElevBehaviour
+}
